MrBot2/requirements.txt
Pillow
SpeechRecognition
art
ascii
asyncpg
buttons
dblpy
discord.py
granitepy
graphviz
jishaku
matplotlib
objgraph
psutil
youtube-dl